fix: destination ID not available when resending packet if route couldn't be found

+++ src/browser_client.rs
@@ -29,7 +29,7 @@ pub struct BrowserClient {
     assembler: Assembler,
     disassembler: Disassembler,
     running: bool,
-    packets_to_send: HashMap<u64, Packet>,
+    packets_to_send: HashMap<u8, Packet>,
     sent_flood_ids: Vec<u64>,
 
     // Specific to browser client
@@ -341,10 +341,10 @@ impl Client for BrowserClient {
         &mut self.running
     }
 
-    fn packets_to_send(&mut self) -> &mut HashMap<u64, Packet> {
+    fn packets_to_send(&mut self) -> &mut HashMap<u8, Packet> {
         &mut self.packets_to_send
     }
-    
+
     fn sent_flood_ids(&mut self) -> &mut Vec<u64> {
         &mut self.sent_flood_ids
     }
